:art: refactor code package directory
