+++ ufs-oss/ufs-oss-aliyun/src/main/java/cc/ccoder/ufs/oss/aliyun/properties/AliyunEncryptKeyProperties.java
@@ -1,4 +1,4 @@
-package cc.ccoder.ufs.oss.properties;
+package cc.ccoder.ufs.oss.aliyun.properties;
 
 import java.util.Map;
